+++ src/Name/ClassName.php
@@ -45,4 +45,9 @@ final class ClassName
     {
         return $this->value;
     }
+
+    public function equals(self $other): bool
+    {
+        return $this->value === $other->value;
+    }
 }