+++ src/Name/ParameterName.php
@@ -37,4 +37,9 @@ final class ParameterName
     {
         return $this->value;
     }
+
+    public function equals(self $other): bool
+    {
+        return $this->value === $other->value;
+    }
 }